create Internal module so that users can get at data constructor if they need to add MTL instance
