+++ stt.cabal
@@ -19,6 +19,7 @@ description:
    combinations of their corresponding monad transformers.
 library
   exposed-modules:     Control.Monad.Trans.ST
+                       Control.Monad.Trans.ST.Internal
   build-depends:       base >= 4.9 && < 4.10
                      , mtl  >= 2.2.1 && < 3
                      , primitive >= 0.6.1 && < 0.7
